Creating working script
This script can map sequences to a genome. It has been tested for the
case of S. pombe. The idea is to use BLAT to map to the genome and only
accept the hits which are exact. So far it does ok. A full discussion of
how it does will be in the notes.