Add script to do pombe analysis
This is my start at making this stuff reproducible and make it easier to
see where all the data came from. In addition, I've starting doing
something that may be odd. That is I am using the downloaded .bed file
and bedtools to extract the sequences of molecules that have genome
coordinates. Depending on the genome used this may not produce the same
sequences as what is in rnacentral, which is odd. This should be checked
as part of the script as well.