Make analysis more or less reproducible
This fixes issues where the bed files were not 0 based but instead 1
based. In addition, I do not exclude a few (4) sequences where the
location information is from silva. This because the location
information is wrong. Also, I convert sequences to RNA so that the blast
search works as expected.