Improve detection of splicing
This does a better job of finding spliced vs unspliced for both hits and
features. This is pretty helpful when examining the results to see why
there are incorrect matches. I've also improved the detection of
identity so it will require 100% identity, not just 0 gaps in the
alignment.

The splicing detection does rely upon the feature of BLAT where it
writes out multiple HSPfragments, so it likely won't work with BLAST.
Not ideal and should be improved. Summarily, the computation of identity
requires a BLAT output.

Also some spelling fixes and the like.