Correct bugs and add new matcher
There were some bugs with how various things were computed. I've
reworked the data to reflect this.

Also I've begun work on a new matcher best-match. This basically takes
as many of the best hits as possible and then will take up to a
specified number of less good hits.