Fix best-match logic
Best match now does more or less the right thing and we take advantage
of it to find the missing coordinates. The idea is to take as many of
the best hits as we can but never take any bad hits. Right now I have it
set to 99% as the minimum allowed difference and 100% query
completeness. This seems to let us match nearly all the unknown
sequences with very few issues compared to 100% match on the known hits.