Begin work on mapping C. elegans
I want to reuse lots of the logic I already have but that is proving
annoying using shell scripts. So I'm moving to a hopefully more reusable
Makefile based approach. This doesn't yet work but does outline what I
think the generic Makefile will be.

The general idea is that most of the logic is the same cross organisms,
the only bit that is different is where to get input
sequences/coordinates and how to do the mapping. I think using a
Makefile I can more or less enforce this.

This is my initial attempt at this. Nothing will work yet but this is
the start of how I will try to move things.