Rework to a Makefile approach
This now can do the pombe mapping using the Makefile. This should be
easy to extend to a new organism like C. elegans. I also reduce the
amount of stuff in the repo so this is easier to clone. This probably
makes it less reproducible as it is possible that the genomes and such
will change without our knowing.