Add more organisms to map
In theory this all works correctly. However, while it can map may of the
unknown sequences it takes too long to map the known ones. This means it
is impossible for me to validate if the selected approach works well.
So I can't really trust the mappings. I will rewrite the mapping logic
to use the cluster and hopefully that will make things efficient enough
that I can actually work with them.