Begin work on script to execute blat remotely
This is a script which will either run blat locally or remotely
depending upon settings in the environment. I think I can slot this into
the normal matcher and have it work to run blat on the server and the
parse results locally. It isn't finished yet as I need to test the
sleep/poll loop for detecting a finished job.